requirements.txt
composio
agno>=2.2.10
streamlit
composio-agno
together
python-docx
requests